+++ Basil/src/Debug/Instrumentor.cpp
@@ -14,14 +14,38 @@ namespace Basil
 	// Begin a session
 	void Instrumentor::beginSession(const std::string& name, const std::string& filePath)
 	{
+		// Close current session before starting new one
+		std::lock_guard lock(mutex);
+		if (currentSession)
+		{
+			if (Log::getEngineLogger())	// Logger may not be initialised before profiling starts
+				LOG_ERROR("Instrumentor::beginSession('{0}') when session '{1}' already open.", name, currentSession->name);
+
+			// End the current session
+			internalEndSession();
+		}
+		
+		// Open file
 		outputStream.open(filePath);
-		writeHeader();
-		currentSession = new InstrumentationSession{ name };
+
+		// Check that file opened
+		if (outputStream.is_open())
+		{
+			currentSession = new InstrumentationSession({ name });
+			writeHeader();
+		}
+		else
+		{
+			if (Log::getEngineLogger())
+				LOG_ERROR("Instrumentor could not open results file '{0}'.", filePath);
+		}
 	}
 
 	// End a session
 	void Instrumentor::endSession()
 	{
+		std::lock_guard lock(mutex);
+		internalEndSession();
 		writeFooter();
 		outputStream.close();
 		delete currentSession;
@@ -32,23 +56,41 @@ namespace Basil
 	// Write the profiling data
 	void Instrumentor::writeProfile(const ProfileResult& result)
 	{
-		if (profileCount++ > 0)
-			outputStream << ",";
+		// Create string stream
+		std::stringstream json;
 
+		// Read in name
 		std::string name = result.name;
 		std::replace(name.begin(), name.end(), '"', '\'');
 
-		outputStream << "{";
-		outputStream << "\"cat\":\"function\",";
-		outputStream << "\"dur\":" << (result.end - result.start) << ',';
-		outputStream << "\"name\":\"" << name << "\",";
-		outputStream << "\"ph\":\"X\",";
-		outputStream << "\"pid\":0,";
-		outputStream << "\"tid\":" << result.threadID << ",";
-		outputStream << "\"ts\":" << result.start;
-		outputStream << "}";
+		// Write data
+		if (profileCount++ > 0)
+			json << ",";
 
-		outputStream.flush();
+		json << "{";
+		json << "\"cat\":\"function\",";
+		json << "\"dur\":" << (result.end - result.start) << ',';
+		json << "\"name\":\"" << name << "\",";
+		json << "\"ph\":\"X\",";
+		json << "\"pid\":0,";
+		json << "\"tid\":" << result.threadID << ",";
+		json << "\"ts\":" << result.start;
+		json << "}";
+
+		// Lock the data and write to file
+		std::lock_guard lock(mutex);
+		if (currentSession)
+		{
+			outputStream << json.str();
+			outputStream.flush();
+		}
+	}
+
+	// Get the instrumentor instance
+	Instrumentor& Instrumentor::get()
+	{
+		static Instrumentor instance;
+		return instance;
 	}
 
 	// Write the header
@@ -65,11 +107,15 @@ namespace Basil
 		outputStream.flush();
 	}
 
-	// Get the instrumentor instance
-	Instrumentor& Instrumentor::get()
+	void Instrumentor::internalEndSession()
 	{
-		static Instrumentor instance;
-		return instance;
+		if (currentSession)
+		{
+			writeFooter();
+			outputStream.close();
+			delete currentSession;
+			currentSession = nullptr;
+		}
 	}
 
 
@@ -98,7 +144,7 @@ namespace Basil
 		long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimePoint).time_since_epoch().count();
 
 		uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
-		Instrumentor::get().writeProfile({ name, start, end, threadID });
+		Instrumentor::get().writeProfile({ name, start, end, std::this_thread::get_id() });
 
 		stopped = true;
 	}